+++ src/handlers/President.java
@@ -17,7 +17,7 @@ public class President extends Approver{
     }
 
     @Override
-    protected int getPurchaseLimit(Type type) {
+    protected double getPurchaseLimit(Type type) {
         return switch (type) {
             case CONSUMABLES -> 1000;
             case CLERICAL -> 2000;
